Refactor automatic tracking gui into a single module
