+++ alignment-software/alignment_software/gui/auto_track.py
@@ -1,7 +1,11 @@
-from tkinter.messagebox import showerror, showinfo
+"""Automic particle tracking step module."""
 
 import os
+import tkinter as tk
+from tkinter import ttk
+from tkinter.messagebox import showerror, showinfo
 import numpy as np
+from .common import NumericSpinbox
 from alignment_software.engine.csv_io import (
     load_marker_csv, write_marker_csv
 )
@@ -9,10 +13,10 @@ from alignment_software.engine.particle_tracking import (
     create_particle_mask,
     particle_search
 )
-from .window_auto_track import AutoTrackWindow
 
 
 MAX_PARTICLES = 13
+INPUT_WIDTH = 10
 
 
 class AutoTrackStep:
@@ -286,3 +290,248 @@ class AutoTrackStep:
     def focus(self):
         """Brings the automatic tracking window to the top."""
         self.auto_track_window.lift()
+
+
+class AutoTrackWindow(tk.Toplevel):
+    """Create the automatic tracking window witth a table and controls."""
+
+    def __init__(self, master, particle_count):
+        """Create the window."""
+        super().__init__(master)
+        self.title("Automatic Detection Window")
+
+        # Adding widgets to the window
+        self.table = ParticleTableFrame(self, particle_count)
+        self.table.grid(column=0, row=0, sticky="we")
+
+        self.track_button = ttk.Button(self, text="Track all selected")
+        self.track_button.grid(column=0, row=1, sticky="we")
+        self.interpolate_button = ttk.Button(self, text="Interpolate selected")
+        self.interpolate_button.grid(column=0, row=2, sticky="we")
+
+        self.properties = ParticlePropertiesFrame(self)
+        self.properties.grid(column=0, row=3, sticky="we")
+
+        self.reset_button = ttk.Button(self, text="Reset all particles")
+        self.reset_button.grid(column=0, row=4, sticky="we")
+
+
+class ParticlePropertiesFrame(tk.Frame):
+    """
+    Frame for tracking properties such as marker radius, search radious and
+    particle color.
+    """
+
+    def __init__(self, master):
+        """Create the frame."""
+        super().__init__(master)
+
+        self.command = None
+
+        self.columnconfigure(0, weight=1)
+        self.columnconfigure(1, weight=1)
+
+        particle_color = tk.LabelFrame(self, text="Particle Color")
+        particle_color.grid(row=0, column=0, sticky="nwse")
+        self.particle_color_var = tk.IntVar(self, 0)
+        black_particle_color = tk.Radiobutton(
+            particle_color, text="Black",
+            variable=self.particle_color_var, value=0
+        )
+        black_particle_color.grid(row=0, column=0)
+        white_particle_color = tk.Radiobutton(
+            particle_color, text="White",
+            variable=self.particle_color_var, value=1
+        )
+        white_particle_color.grid(row=0, column=1)
+
+        search_areas_frame = tk.LabelFrame(self, text="Shift Search Areas")
+        search_areas_frame.grid(row=0, column=1, sticky="nwse")
+
+        marker_radius_label = ttk.Label(
+            search_areas_frame, text="Marker radius (pixel)"
+        )
+        marker_radius_label.grid(row=0, column=0)
+        self.marker_radius_input = NumericSpinbox(
+            search_areas_frame, width=INPUT_WIDTH,
+            value_default=80, value_range=[1, 999],
+            command=self.update
+        )
+        self.marker_radius_input.grid(row=0, column=1)
+
+        search_area_width_label = ttk.Label(
+            search_areas_frame, text="Search width (pixel)"
+        )
+        search_area_width_label.grid(row=1, column=0)
+        self.search_area_width_input = NumericSpinbox(
+            search_areas_frame, width=INPUT_WIDTH,
+            value_default=80, value_range=[1, 999],
+            command=self.update
+        )
+        self.search_area_width_input.grid(row=1, column=1)
+
+        search_area_height_label = ttk.Label(
+            search_areas_frame, text="Search height (pixel)"
+        )
+        search_area_height_label.grid(row=2, column=0)
+        self.search_area_height_input = NumericSpinbox(
+            search_areas_frame, width=INPUT_WIDTH,
+            value_default=80, value_range=[1, 999],
+            command=self.update
+        )
+        self.search_area_height_input.grid(row=2, column=1)
+
+    def set_command(self, command):
+        """Set the command to be called when properties update."""
+        self.command = command
+
+    def update(self):
+        """Call the command when properties are updated."""
+        if self.command is not None:
+            self.command()
+
+    def set_properties(self, properties):
+        """Set the properties from dictionary."""
+        self.search_area_width_input.set(properties["search_size"][0])
+        self.search_area_height_input.set(properties["search_size"][1])
+        self.marker_radius_input.set(properties["marker_radius"])
+
+    def get_properties(self):
+        """Get properties as dictionary."""
+        return {
+            "search_size": (
+                self.search_area_width_input.get(),
+                self.search_area_height_input.get()
+            ),
+            "marker_radius": self.marker_radius_input.get()
+        }
+
+    def get_invert_particle_color(self):
+        """
+        Get whether the particle color is inverted.
+        Default is dark, inverted is bright.
+        """
+        return self.particle_color_var.get() == 1
+
+
+
+class ParticleTableFrame(tk.Frame):
+    """Create a table which shows tracking locations, and frame ranges."""
+
+    def __init__(self, master, particle_count):
+        """Create the table frame with a given number of particles."""
+        super().__init__(master)
+
+        table_header = ["X", "Y", "IM1", "IM2"]
+        for i, header in enumerate(table_header):
+            header = tk.Label(self, text=header)
+            header.grid(row=0, column=i+2)
+
+        self.mark_end_command = None
+        self.reset_command = None
+        self.particle_select_var = tk.IntVar(self, 0)
+        self.data_vars = [[None] * 5 for i in range(particle_count)]
+        self.track_vars = []
+
+        for i in range(particle_count):
+
+            particle_select = tk.Radiobutton(
+                self, text=f"{i+1}", value=i,
+                variable=self.particle_select_var
+            )
+            particle_select.grid(row=i+1, column=0)
+
+            data_var = tk.StringVar(self, "o")
+            label = tk.Label(self, textvariable=data_var)
+            label.grid(row=i+1, column=1, sticky="nswe")
+            self.data_vars[i][0] = data_var
+
+            for c in range(4):
+                data_var = tk.StringVar(self, "-")
+                label = tk.Label(
+                    self, textvariable=data_var,
+                    width=8, bd=1, relief="solid"
+                )
+                label.grid(row=i+1, column=c+2, sticky="nswe")
+                self.data_vars[i][c+1] = data_var
+
+            end_button = tk.Button(
+                self, text="Mark End",
+                command=lambda particle_index=i: self.mark_end(particle_index)
+            )
+            end_button.grid(row=i+1, column=6)
+
+            reset_button = tk.Button(
+                self, text="Reset",
+                command=lambda particle_index=i: self.reset(particle_index)
+            )
+            reset_button.grid(row=i+1, column=7)
+
+            track_var = tk.BooleanVar(self, False)
+            self.track_vars.append(track_var)
+            track_checkbox = tk.Checkbutton(
+                self, text="Select", variable=track_var
+            )
+            track_checkbox.grid(row=i+1, column=8)
+
+    def update_data(
+        self, particle_positions,
+        tracking_positions, tracking_start_frames, tracking_end_frames
+    ):
+        """Update the contents of the table."""
+        for i, position in enumerate(tracking_positions):
+            status = particle_positions.get_status(i)
+            if status == "complete":
+                icon = "●"
+            elif status == "partial":
+                icon = "◒"
+            else:
+                icon = "○"
+            self.data_vars[i][0].set(icon)
+            if position is not None:
+                self.data_vars[i][1].set(position[0])
+                self.data_vars[i][2].set(position[1])
+            else:
+                self.data_vars[i][1].set("-")
+                self.data_vars[i][2].set("-")
+            self.data_vars[i][3].set(tracking_start_frames[i]+1)
+            self.data_vars[i][4].set(tracking_end_frames[i]+1)
+
+    def get_selected_particle(self):
+        """Get the index of the particle selected by the radio button"""
+        return self.particle_select_var.get()
+
+    def get_tracked_particles(self):
+        """Get particles selected for tracking."""
+        tracked = []
+        for i, variable in enumerate(self.track_vars):
+            if variable.get():
+                tracked.append(i)
+        return tracked
+
+    def enable_tracking(self, particle_index):
+        """Selects a particle for tracking."""
+        self.track_vars[particle_index].set(True)
+
+    def disable_tracking(self, particle_index):
+        """Deselects a particle for tracking."""
+        self.track_vars[particle_index].set(False)
+
+    def set_mark_end_command(self, command):
+        """Sets a command that gets called when mark end is clicked."""
+        self.mark_end_command = command
+
+    def mark_end(self, particle_index):
+        """Called when a mark button is clicked with a given index."""
+        if self.mark_end_command is not None:
+            self.mark_end_command(particle_index)
+
+    def set_reset_command(self, command):
+        """Sets a command that gets called when reset is clicked."""
+        self.reset_command = command
+
+    def reset(self, particle_index):
+        """Called when a reset button is clicked with a given index."""
+        self.disable_tracking(particle_index)
+        if self.reset_command is not None:
+            self.reset_command(particle_index)
