+++ alignment-software/alignment_software/gui/window_main.py
@@ -10,7 +10,7 @@ from .contrast.step_contrast import ContrastStep
 from .loading.step_loading import LoadingStep
 from .transform.step_transform import TransformStep
 from .coarse_align.step_coarse_align import CoarseAlignStep
-from .auto_track.step_auto_track import AutoTrackStep
+from .auto_track import AutoTrackStep
 from .optimization.step_optimization import OptimizationStep
 from .frame_steps import StepsFrame
 from .frame_image import ImageFrame
